bug fix: fetch format and size in case of read from another partition
The overall flow in order to fetch the compression format and
uncompressed block size is as follows:
1. call requestCompFormatSize() when a connection to fog of the local
partition is made in order to find mbid locations for read.
2. If the read has to be made locally. Got lucky, since we have the
format and size already. Can proceed with the read from edge directly.
3. If the read has to be done through another fog(i.e in another
partition). Then we call requestCompFormat() on the other fog.

NOTE:  the previous implementation of write, there was no compression and hence
there was no need to have the values of compression format and uncompressed block size
before hand in order to perform a read

This implementation has minimal overheads since the required block metadata
is only fetched when a connection has already been made to a
fog(either local or of another partition).

Another alternative to get the required metadata in one call would
involve search on multiple fogs via their bloom filters. This would be
quite expensive.